Se extrae la logica de axios de InfoGestion
